requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-filter==25.2
drf-spectacular==0.29.0

celery==5.3.6
django-celery-results==2.6.0
redis==5.0.4

gunicorn==21.2.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1

pillow==10.4.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
tzdata==2024.1

django-redis==5.4.0
python-slugify==8.0.4